requirements.txt
fastapi
google-generativeai
pandas
python-dotenv
uvicorn[standard]
deepseek
openai